% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_overview.R
\docType{package}
\name{davtools}
\alias{davtools}
\title{davtools: A package to improve my life}
\description{
The package davtools provides one function and a dataset:
print_bf, a function to print Bayes Factors in Markdown files
res_rt, a data frame with simulated results of an experiment.
}
\section{Functions}{

The functions \code{\link{print_bf}} can be used to print Bayes Factors in Markdown files
}

\section{Datasets}{

The datasets are used to test and show different functions in this package
}

