% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{Compute grid of unique combinations}
\usage{
expand_grid_unique(x, y, include_eq = FALSE)
}
\arguments{
\item{x}{variable}

\item{y}{variable}

\item{include_eq}{include eq combinations TRUE/FALSE}
}
\value{
RMSE
}
\description{
Function to make a grid of uniqe factor combinations, i.e. ignoring order of factors.
}
\examples{
expand_grid_unique(c("a","b","c"),c("a","b","c"))

}
\author{
David Izydorczyk
}
