% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenlist.R
\name{flattenlist}
\alias{flattenlist}
\title{Flatten a highly nested list}
\usage{
flattenlist(x)
}
\arguments{
\item{x}{list}
}
\value{
simple list
}
\description{
Function to flatten a highly nested list as returned by \code{foreach()} into a simple list
}
\examples{
flattenlist(list(1,list(2,list(3,list(4,5)))))

}
\author{
David Izydorczyk
}
