% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_readme.R
\name{make_df_readme}
\alias{make_df_readme}
\title{Make Readme of Data.frame}
\usage{
make_df_readme(
  df,
  desc,
  info = NULL,
  file = "readme.txt",
  add_examples = TRUE,
  digits = 2
)
}
\arguments{
\item{df}{data.frame to generate readme for}

\item{desc}{vector of descriptions for all variables in df}

\item{info}{additional info you might want to add before the data.frame, prints also the dimensions of the data.frame}

\item{file}{name and path of the output file}

\item{add_examples}{if examples of data in the variables should be added to the readme}

\item{digits}{number of digits used in the examples column (if numeric variable)}
}
\value{
a .txt file with a markdown summary table of your data.frame
}
\description{
Prepares a summary of a data.frame in form of a markdown table, including the name of the columns, a short description, and example values for each variable.
}
\examples{
\dontrun{
df = data.frame("ID"   = 1:50,
                "cond" = sample(c("A","B"),50,replace=TRUE),
                "rt"   = rnorm(50,500,40))

desc = c("unique numeric participant ID",
         "condition [A,B]",
         "average reaction time in ms")

make_df_readme(df   = df,
               desc = desc,
               info = "Data of the experiment Hello World",
               add_examples=TRUE)
}




}
\author{
David Izydorczyk
}
