% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Make Print mean and sd  in markdown}
\usage{
mean_sd(x, digits = 2, gt1 = TRUE)
}
\arguments{
\item{x}{numerical variable}

\item{digits}{number of digits used in the examples column (if numeric variable)}

\item{gt1}{variable can take on values greater than 1}
}
\value{
a character string with the pattern "Mean (SD)"
}
\description{
Convinience function for printing \code{mean},  \code{sd} (and range) in markdown
}
\examples{
mean_sd(rnorm(100))
"0.10 (1.01)"
mean_sd(runif(100,0,1),gt1=FALSE)
".49 (.29)"

}
\author{
David Izydorczyk
}
