% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bf.R
\name{print_bf}
\alias{print_bf}
\title{Report Bayes Factors in Manuscripts}
\usage{
print_bf(bf)
}
\arguments{
\item{bf}{The Bayes Factor which should be reported}
}
\value{
Returns the Bayes Factor as string. Bayes factors greater than 1000 are abbreviated.
}
\description{
\code{print_bf} returns the Bayes Factor according to current reporting standards.
}
\examples{
print_bf(10)
"10"
print_bf(30.32552)
"30.32"
print_bf(5235)
"> 1000"
}
\author{
David Izydorczyk
}
