% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbern.R
\name{rbern}
\alias{rbern}
\title{Generate Random Bernoulli Distributed Samples}
\usage{
rbern(n, prob)
}
\arguments{
\item{n}{number of samples}

\item{prob}{probability of X = 1}
}
\value{
rbern
}
\description{
Function to compute random Bernoulli distributed samples. Wrapper around \code{rbinom}.
}
\examples{
rbern(10,0.5)

}
\author{
David Izydorczyk
}
