% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{res_rt}
\alias{res_rt}
\title{Example data of a reaction time experiment}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{ID}{the ID of the person}
\item{cond}{the condition the person is in}
\item{rt}{the reaction time in ms}
}
}
\source{
simulated
}
\usage{
res_rt
}
\description{
The dataset containing the reaction times of 50 fictitious participants in two conditions.
}
\keyword{datasets}
