% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Compute Root-Mean-Squared Error}
\usage{
RMSE(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\value{
RMSE
}
\description{
Function to compute the root-mean-squared-error (RMSE) between two numeric variables
}
\examples{
RMSE(rnorm(100),rnorm(100))

}
\author{
David Izydorczyk
}
