% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save results}
\usage{
save_results(results, path, name)
}
\arguments{
\item{results}{results object to save}

\item{path}{path were to save the object}

\item{name}{file name}
}
\value{
Nothing
}
\description{
Function to save an R object as .Rdata with the current date automatically appended to the file name
}
\examples{
install_n_load_packages()

}
\author{
David Izydorczyk
}
