% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_to_tau.R
\name{sigma_tau}
\alias{sigma_tau}
\title{Convert sigma to tau (i.e. precision in JAGS)}
\usage{
sigma_tau(tau = NA, sigma = NA)
}
\arguments{
\item{tau}{precision}

\item{sigma}{standard deviation}
}
\value{
sigma or tau
}
\description{
Function to transform sigma to tau (i.e. precision in JAGS) or the other way around
}
\examples{
sigma_tau(tau   = 0.5)
sigma_tau(sigma = 4)

}
\author{
David Izydorczyk
}
