% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_overview.R
\docType{package}
\name{davtools}
\alias{davtools}
\title{davtools: A package to improve my life}
\description{
The package davtools provides several function and one simulated dataset.
}
\section{Functions}{

\itemize{
\item \code{\link{print_bf}}:  print Bayes Factors in Markdown files
\item \code{\link{make_df_readme}}:  create summary of a data.frame in form of a markdown table
\item \code{\link{mean_sd}}:  print mean and standarddeviation of a variable in Markdown files
\item \code{\link{mean_sd}}: compute the root-mean-squared-error (RMSE) between two numeric variables
\item \code{\link{expand_grid_unique}}:  make a grid of uniqe factor combinations, i.e. ignoring order of factors
\item \code{\link{sigma_tau}}:  transform sigma to tau (i.e. precision in JAGS) or the other way around
\item \code{\link{flattenlist}}:  flatten a highly nested list as returned by \code{foreach()} into a simple list
\item \code{\link{install_n_load_packages}}:  load packages and install missing packages if necessary
\item \code{\link{save_results}}:  save an R object as .Rdata with the current date automatically appended to the file name
\item \code{\link{rbern}}: generate random Bernoulli distributed samples
}
}

\section{Datasets}{

The datase  \code{\link{res_rt}} is used to test and show different functions in this package
}

